Renamed get_weather_data.py to Data.py
Renamed get_weather_data.py to Data.py. Updated import statements where needed to reflect this change.

Added the Binarize function to Data.py to binarize weather data.
